+++ test/flycheck-test.el
@@ -4089,6 +4089,12 @@ Why not:
    '(2 nil error "An argument definition must end with a newline."
        :checker terraform)))
 
+(flycheck-ert-def-checker-test terraform-tflint terraform nil
+  (flycheck-ert-should-syntax-check
+   "language/terraform/error.tf" 'terraform-mode
+   '(3 nil error "\"t1.2xlarge\" is invalid instance type."
+       :id "aws_instance_invalid_type" :checker terraform-tflint)))
+
 (flycheck-ert-def-checker-test markdown-markdownlint-cli markdown nil
   (flycheck-ert-should-syntax-check
    "language/markdown.md" 'markdown-mode